python/requirements.txt
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
# NepseUnofficialApi – real-time data from nepalstock.com
nepse>=1.1.0
